update: DateWrapper for theme "date-and-time"
